backend/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["kylin <260987762@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
mongoengine = "^0.28.2"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
drf-spectacular = "^0.27.1"
marshmallow-mongoengine = "^0.31.2"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
pydantic = "^2.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
